requirements.txt
# Core API and Web Framework
fastapi==0.115.0
uvicorn==0.30.0

# Face Recognition and Image Processing
face_recognition==1.3.0
face_recognition_models==0.3.0
dlib==19.24.6
Pillow==10.4.0

# Vector Search
faiss-cpu==1.13.0

# AWS Integration
boto3==1.35.0

# Database
sqlite-utils==3.36.0

# Utilities
numpy==1.26.4
requests==2.32.3

# UI Apps
tk==0.1.0  # For Tkinter-based UI